+++ app.rb
@@ -1,79 +1,30 @@
-require './person'
-require './teacher'
-require './student'
-require './book'
-require './rental'
+require './menu/menu'
+require './functions/data'
+require './functions/add_person'
+require './functions/add_book'
+require './functions/add_rental'
+require './functions/person_list'
+require './functions/book_list'
+require './functions/rental_list'
 
-class Data
-  attr_accessor :person, :book, :rentals
-
-  def initialize
-    @person = []
-    @book = []
-    @rentals = []
-  end
-
-  def add_student(age, name, _permission, classroom)
-    student_data = Student.new(classroom, age, name)
-    @person << student_data
-    puts 'Person created successfully'
-  end
-
-  def add_teacher(specialization, age, name)
-    teacher_data = Teacher.new(specialization, age, name)
-    @person << teacher_data
-    puts 'Person created successfully'
-  end
-
-  def add_book(title, author)
-    book_data = Book.new(title, author)
-    @book << book_data
-    puts 'Book created successfully'
-  end
-
-  def add_rental(date, book, person)
-    @book.filter_map do |info|
-      @title = info.title if info.title == book
-      @author = info.author if info.title == book
-    end
-    book_data = Book.new(@title, @author)
-
-    @person.filter_map do |info|
-      @age = info.age if info.name == person
-      @name = info.name if info.name == person
-    end
-    person_data = Person.new(@age, @name)
-
-    rental_data = Rental.new(date, book_data, person_data)
-    @rentals << rental_data
-    puts 'Rental created successfully'
-  end
-
-  def list(value)
-    case value
-    when 'book'
-      if @book.empty?
-        puts 'There are no books'
-      else
-        (@book.map.with_index do |book, index|
-          puts "#{index}) Title: #{book.title}, Author: #{book.author}"
-        end)
+def app
+  data = Data.new
+  loop do
+    load_menu
+    if valid_input?(@input)
+      case @input
+      when '1' then BookList.new.book_list(data.book)
+      when '2' then PersonList.new.person_list(data.person)
+      when '3' then AddPerson.new.add_person(data.person)
+      when '4' then AddBook.new.add_book(data.book)
+      when '5' then AddRental.new.add_rental(data.book, data.person, data.rentals)
+      when '6' then RentalList.new.rental_list(data.rentals, data.person)
+      when '7'
+        puts 'Thank you for using the app!'
+        break
       end
-    when 'people'
-      if @person.empty?
-        puts 'There are no records'
-      else
-        (@person.map.with_index do |val, index|
-          puts "#{index}) Name: #{val.name}, ID: #{val.id}, Age: #{val.age}"
-        end)
-      end
+    else
+      puts "option #{@input} is invalid, please try again"
     end
-  end
-
-  def rental_list(value)
-    puts "Name: #{value}"
-    (@rentals.filter_map do |res|
-      puts "Date: #{res.date}, Book '#{res.book.title}' by #{res.book.author}" if res.person.name == value
-    end)
   end
 end
