+++ correctName/trimmer.rb
@@ -1,0 +1,7 @@
+require_relative './decorator'
+
+class TrimmerDecorator < Decorator
+  def correct_name
+    @nameable.correct_name.strip.slice(0, 10)
+  end
+end
