+++ functions/data.rb
@@ -1,0 +1,9 @@
+class Data
+  attr_accessor :person, :book, :rentals
+
+  def initialize
+    @person = []
+    @book = []
+    @rentals = []
+  end
+end
