+++ menu/valid.rb
@@ -1,0 +1,7 @@
+def valid_input?(input)
+  %w[1 2 3 4 5 6 7].include?(input)
+end
+
+def valid_person?(selection)
+  %w[1 2].include?(selection)
+end
