+++ school/rental.rb
